Add files via upload
did not add everything to the signin/out button (takeactionlistener) like how to choose what teacher and what item to sign back in or sign out. Also did not have time to do the restock tab, will do monday and tuesday. Only thing left is to add functionallity to all of the buttons and to make the settings and help tab, which i will do with some help