Add files via upload
This is a project i did last year that had classes that accessed private variables even though they were outside of the scope. Look at ContactGui mainly because that is the gui. Also even though the actionlisteners are in the same .java file as the main gui, they are seperate classes.