Add files via upload
The code i was trying to integrate, there are some changes in the sign in/out button listener but the borrowing part doesn't have any list of teachers imported because i don't know where the teachers array is. Also some changes were made to the item class for it to work.